libcamera: media_device: Handle ancillary links in populateLinks()
The populateLinks() function can't currently handle ancillary links
which causes an error to be thrown in process() when the MediaObject
cannot be cast to a MediaPad.

Add explicit handling for the different link types, creating either
pad-2-pad or entity-2-entity versions of a MediaLink based on the
link type in the link's flags.

Signed-off-by: Daniel Scally <djrscally@gmail.com>