libcamera: ipu3-cio2: Discover VCMs through ancillary links
Rather than attempting to discover VCMs via model name matching we
should follow the ancillary links that define the relationship
between the two devices to locate any entities linked to the sensor.
Where we have linked entities with the function MEDIA_ENT_F_LENS we
can then create an instance of CameraLens to represent it.

Signed-off-by: Daniel Scally <djrscally@gmail.com>